scsi: fix occasional bogus elapsed time that exceeds timeout in /proc/scsi/sg/debug
A race condition was found in sg_proc_debug_helper(). It was observed on
a system using an IBM LTO-9 SAS Tape Drive (ULTRIUM-TD9) and monitoring
/proc/scsi/sg/debug every second. A very large elapsed time would
sometimes appear.

The likely cause is a race between the reading of hp->duration in
sg_proc_debug_helper() and request completion in sg_rq_end_io().
The hp->duration member variable may hold either of two types of
information:
\#1 - The start time of the request. This value is present while
      the request is not yet finished.

\#2 - The total execution time of the request (end_time - start_time).

If sg_proc_debug_helper() executes *after* the value of hp->duration
was changed from \#1 to \#2, but *before* srp->done is set to 1 in
sg_rq_end_io(), a fresh timestamp is taken in the else branch and the
elpsed time (value type \#2) is substracted from a timestamp, which
cannot yield a valid elapsed time, which is also a type \#2 value.

The fix involves adding a new member variable time_started to the
sg_request struct that shall contain the timestamp of the request
starting in milliseconds. This value shall not change and shall
be used instead of the duration member variable of the sg_io_hdr_t
struct. That member variable shall now be initialized to 0 once the
request is sent in sg_common_write() and will change its value to
the real elapsed time once the request is resolved in sg_rq_end_io().

The issue was reproduced successfully. A patched kernel was built and
the race condition could not be observed anymore after the application
of this patch. The reproducer C program can be viewed here:
https://github.com/MichaelRabek/linux-tests/blob/master/drivers/scsi/sg/sg_race_trigger.c

Tested-by: Changhui Zhong <czhong@redhat.com>
Signed-off-by: Michal Rábek <mrabek@redhat.com>