scsi: fix occasional bogus elapsed time that exceeds timeout
A race condition was found in sg_proc_debug_helper(). It was observed on
a system using an IBM LTO-9 SAS Tape Drive (ULTRIUM-TD9) and monitoring
/proc/scsi/sg/debug every second. A very large elapsed time would
sometimes appear.

The likely cause is a race between the reading of hp->duration in
sg_proc_debug_helper() and request completion in sg_rq_end_io().
The hp->duration member variable may hold either of two types of
information:
 #1 - The start time of the request. This value is present while
      the request is not yet finished.

 #2 - The total execution time of the request (end_time - start_time).

If sg_proc_debug_helper() executes *after* the value of hp->duration
was changed from #1 to #2, but *before* srp->done is set to 1 in
sg_rq_end_io(), a fresh timestamp is taken in the else branch and the
elapsed time (value type #2) is subtracted from a timestamp, which
cannot yield a valid elapsed time, which is also a type #2 value.

The fix involves making sure hp->duration is only read once in the
else branch. Even if the race condition does occur and hp->duration
ends up containing the #2 type value, ms - hp->duration is going to
produce a large number; a time much larger than the timeout of the
request. In this case, 0 shall be printed.

The issue was reproduced successfully on an IBM ULTRIUM-HH9 tape drive.
A patched kernel was built, and the race condition could not be observed
anymore after the application of this patch. A reproducer C program
utilising the scsi_debug module was also built by Changhui Zhong and
can be viewed here:
https://github.com/MichaelRabek/linux-tests/blob/master/drivers/scsi/sg/sg_race_trigger.c

Reviewed-by: Tomas Henzl <thenzl@redhat.com>
Tested-by: Changhui Zhong <czhong@redhat.com>
Signed-off-by: Michal Rábek <mrabek@redhat.com>